Fix a few small issues with query logging.
git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk@16070 18c085ea-50e0-402c-830e-de6fd14e8384

+++ src/gui/QvisXRayImageQueryWidget.C
@@ -319,7 +319,7 @@ QvisXRayImageQueryWidget::GetQueryParameters(MapNode &params)
     if (noerrors)
     {
         params["output_type"] = imageFormat->currentIndex();
-        params["divide_emis_by_absorb"] = divideFlag->isChecked();
+        params["divide_emis_by_absorb"] = (int)divideFlag->isChecked();
         params["origin"] = origin;
         params["theta"] = t;
         params["phi"] = p;
